requirements.txt
aiofiles==24.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
carvekit==4.1.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
fsspec==2024.12.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pillow_heif==0.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
python-multipart==0.0.20
requests==2.32.3
Rx==3.2.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tkinterdnd2==0.4.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
